[security] Fix CVE-2011-5000
The ssh_gssapi_parse_ename function in gss-serv.c in OpenSSH 5.8
and earlier, when gssapi-with-mic authentication is enabled, allows
remote authenticated users to cause a denial of service (memory
consumption) via a large value in a certain length field.
NOTE: there may be limited scenarios in which this issue is relevant.